python-security-lab-demo/requirements.txt
pyyaml==5.3.1  # vulnerable on purpose
